Move error handler views to their own file
