+++ booktracker/urls.py
@@ -25,7 +25,7 @@ urlpatterns = [
     # path('people/add', ),
 ]
 
-handler400 = "booktracker.views.error400"
-handler403 = "booktracker.views.error403"
-handler404 = "booktracker.views.error404"
-handler500 = "booktracker.views.error500"
+handler400 = "booktracker.errorviews.error400"
+handler403 = "booktracker.errorviews.error403"
+handler404 = "booktracker.errorviews.error404"
+handler500 = "booktracker.errorviews.error500"
